Use Double.parseDouble for squeezebox volume and then cast it to an int otherwise a non integer value causes a crash

+++ bundles/io/org.openhab.io.squeezeserver/src/main/java/org/openhab/io/squeezeserver/SqueezeServer.java
@@ -495,7 +495,7 @@ public class SqueezeServer implements ManagedService {
 				// Parameter Volume
 				else if (messagePart.startsWith("mixer%20volume%3A")) {
 					String value = messagePart.substring("mixer%20volume%3A".length());
-					player.setVolume(Integer.parseInt(value));
+					player.setVolume((int) Double.parseDouble(value));
 				}
 				// Parameter Mode
 				else if (messagePart.startsWith("mode%3A")) {
